# src/swagger.yaml
openapi: 3.0.0
info:
  title: Home Like
  description: API documentation for HomeLike API
  version: 0.0.1
servers:
  - url: http://localhost:3010
    description: local
paths:
...
  /api/v1/apartment:
    post:
      summary: create appartment
      description: create appartment
      tags: 
        - Apartment
      requestBody:
        $ref: '#/components/requestBodies/CreateApartment'
      responses:
        '200':
          description: Success
          content:
            appliaction/json:
              schema:
                $ref: '#/components/schemas/Apartment'
        '401':
            $ref: '#/components/responses/UnauthorizedError'
        '406':
            $ref: '#/components/responses/InvalidRequest'
    get:
      summary: search apartments
      # desciption: search aparmtments by city, state, location
      tags:
        - Apartment
      parameters:
        - $ref: '#/components/parameters/searchQuery'
        - $ref: '#/components/parameters/city'
        - $ref: '#/components/parameters/state'
        - $ref: '#/components/parameters/country'
        - $ref: '#/components/parameters/withinLat'
        - $ref: '#/components/parameters/withinLong'
      responses:
        '200':
          description: Apartment list
          content:
            appliaction/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/Apartment'
  /api/v1/apartment/{id}:  
    get:
      summary: get apartment by id
      description: get apartment by id
      tags:
        - Apartment
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
            description: Success
            content:
              appliaction/json:
                schema:
                  $ref: '#/components/schemas/Apartment'
        '401':
            $ref: '#/components/responses/UnauthorizedError'
        '406':
            $ref: '#/components/responses/InvalidRequest'
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    withinLat:
      name: withInLat
      in: query
      schema:
        type: number
        example: 1.233
    searchQuery:
      name: searchQuery
      in: query
      schema:
        type: string
        example: uae
    withinLong:
      name: withInLong
      in: query
      schema:
        type: number
        example: 13.233
    country:
      name: country
      in: query
      schema:
        type: string
        example: DE
    state:
      name: state
      in: query
      schema:
        type: string
        example: HE
    city:
      name: city
      in: query
      schema:
        type: string
        example: FrankFurt
    id:
      name: id
      description: id
      in: path
      required: true
      schema:
         $ref: '#/components/schemas/ObjectId'
...
  schemas:
    Apartment:
      title: Apartment
      type: object
      properties:
        title:
          type: string
          example: Cosy Apartments with kitchen - stylish & comfortable wit
        description:
          type: string
          example: We would like to welcome you to Frankfurt West, a central and beautiful district that is currently undergoing a very exciting development. 150 apartments from 20 to 40 sqm are waiting for you. Some of them with balcony or terrace
        user:
          $ref: '#/components/schemas/ObjectId'
        totalRooms:
          type: number
          example: 2
        apartmentType:
          type: string
          example: guest
        code:
          type: string
          example: DE-12345
        amenities:
          type: object
          properties:
            tv: 
              type: boolean
            heating:
              type: boolean
            towelAndSheets:
              type: boolean
            washingMachine:
              type: boolean
            workDesk:
              type: boolean
        location:
          type: object
          properties:
            streetAddress:
              type: string
              example: Solmsstraße , 60486 ,Frankfurt am Main, Deutschland
            city:
              type: string
              example: Frankfurt am Main
            state:
              type: string
              example: HE
            country:
              type: string
              example: DE
            loc:
              type: object
              properties:
                coordinates:
                  type: array
                  items:
                    type: number
                    example: [1.2222,2.3333]
...
    ObjectId:
      description: Mongoose object id
      example: 5ef052ae6a952a535d3a695f
      type: string
      pattern: ^[a-f\d]{24}$
...
  responses:
    UnauthorizedError:
      description: Unauthorised error
      content:
        application/json:
          schema: 
            type: object
            properties:
              statusCode:
                type: integer
                example: 401
              code:
                type: string
                example: UnauthorizedError
              message:
                type: string
                example: Unauthorized
    InvalidRequest:
      description: Invalid request
      content:
        application/json:
          schema: 
            type: object
            properties:
              code:
                type: string
                example: InvalidRequest
              message:
                type: string
                example: Invalid request
              title:
                type: string
                example: Invalid
...
  requestBodies:
...
    CreateApartment:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              title:
                type: string
                example: Cosy Apartments with kitchen - stylish & comfortable wit
              apartmentType:
                  example: studio
              description:
                type: string
                example: '150 apartments from 20 to 40 sqm are waiting for you. Some of them with balcony or terrace. The “Salone”, our community area, is at your disposal if you don’t want to work in the apartment or if you just feel like some company.'
              country:
                type: string
                example: DE
              state:
                type: string
                example: HE
              city:
                type: string
                example: 25222
              lat:
                type: number
                example: 50.11552000
              long:
                type: number
                example: 8.68417000